+++ packages/orq-rc/src/models/components/conversationwithmessagesresponse.ts
@@ -122,7 +122,7 @@ export const ConversationWithMessagesResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kgpkw6s9cj66ejananhr9rfa"),
+  _id: z.string().default("conv_01kgprqrdh8zjbtqbbk25eb5aj"),
   kind: ConversationWithMessagesResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),